// src/types/adguard.d.ts
export interface components {
    schemas: {
// ...
        Filter: {
            enabled: boolean;
            /**
             * Format: int64
             * @example 1234
             */
            id: number;
            /**
             * Format: date-time
             * @example 2018-10-30T12:18:57+03:00
             */
            last_updated?: string;
            /** @example AdGuard Simplified Domain Names filter */
            name: string;
            /**
             * Format: uint32
             * @example 5912
             */
            rules_count: number;
            /** @example https://adguardteam.github.io/AdGuardSDNSFilter/Filters/filter.txt
             *      */
            url: string;
        };
// ...
    };
// ...
}